refactor(projects): projects stored in json
