+++ src/projects/parser.ts
@@ -1,0 +1,7 @@
+export type Project = {
+    name: string,
+    link: string,
+    image: string,
+    alt: string,
+    content: string
+}
